# backend/src/swagger/docs.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Mood Movie Finder",
    "description": "Recommends movies based on the current weather"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/recommend/{city}": {
      "get": {
        "summary": "Get recommendations by city's weather",
        "description": "Returns the current weather and recommended movies.",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the city"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WeatherCondition": {
        "type": "string",
        "enum": [
          "clear sky",
          "few clouds",
          "scattered clouds",
          "broken clouds",
          "shower rain",
          "rain",
          "thunderstorm",
          "snow",
          "mist"
        ]
      },
      "Movie": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "Inception" },
          "overview": {
            "type": "string",
            "example": "A mind-bending thriller about dream invasion."
          },
          "posterPath": {
            "type": "string",
            "example": "/1E5baAaEse26fej7uHcjOgEE2t2.jpg"
          },
          "rating": { "type": "number", "example": 4.9 }
        }
      },
      "RecommendResponse": {
        "type": "object",
        "properties": {
          "weather": {
            "$ref": "#/components/schemas/WeatherCondition"
          },
          "movies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Movie"
            }
          }
        }
      }
    }
  }
}